redisigned rule app of api
