+++ snort3_monitor/monitor/models.py
@@ -1,9 +1,9 @@
 from django.db import models
-from django.shortcuts import get_object_or_404
+from rule.models import Rule
 
 
 class Event(models.Model):
-    rule = models.ForeignKey('Rule', on_delete=models.CASCADE)
+    rule = models.ForeignKey(Rule, on_delete=models.CASCADE)
     timestamp = models.DateTimeField()
     src_addr = models.CharField(max_length=128)
     src_port = models.IntegerField(null=True, blank=True)
@@ -11,21 +11,6 @@ class Event(models.Model):
     dst_port = models.IntegerField(null=True, blank=True)
     proto = models.CharField(max_length=128)
     mark_as_deleted = models.BooleanField(default=False)
-
-
-class Rule(models.Model):
-    sid = models.IntegerField()
-    gid = models.IntegerField()
-    rev = models.IntegerField()
-    action = models.CharField(max_length=50)
-    message = models.TextField()
-    data_json = models.JSONField()
-
-    @staticmethod
-    def get_rule(sid: int, rev: int, gid: int) -> 'Rule':
-        """Checking for existing of concrete rule"""
-        rule = get_object_or_404(Rule, sid=sid, rev=rev, gid=gid)
-        return rule
 
 
 class Request(models.Model):
