+++ snort3_monitor/rule/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from rule.views import RuleCreate, RuleListView
+
+
+urlpatterns = [
+    path("", RuleListView.as_view(), name='rules-list'),
+    path("update/", RuleCreate.as_view(), name='rules-create'),
+]
