+++ snort3_monitor/script_rules.py
@@ -8,7 +8,7 @@ from django.http import Http404
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "snort3_monitor.settings")
 django.setup()
-from monitor.models import Rule
+from rule.models import Rule
 
 
 logger = logging.getLogger('rules')
