+++ snort3_monitor/snort3_monitor/settings.py
@@ -38,8 +38,9 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "monitor.apps.MonitorConfig",
+    "rule.apps.RuleConfig",
     "rest_framework",
-    'corsheaders',
+    "corsheaders",
 ]
 
 MIDDLEWARE = [
