+++ snort3_monitor/snort3_monitor/urls.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
-from django.urls import path
+from django.urls import path, include
 
-from monitor.views import EventListUpdate, EventCountList, RequestList, RuleCreate, RuleListView
+from monitor.views import EventListUpdate, EventCountList, RequestList
 
 
 handler404 = 'monitor.views.error404'
@@ -11,6 +11,5 @@ urlpatterns = [
     path("api/v1/events/", EventListUpdate.as_view(), name='event-list-update'),
     path("api/v1/events/count/", EventCountList.as_view(), name='event-count-list'),
     path("api/v1/requests-log/", RequestList.as_view(), name='request-list'),
-    path("api/v1/rules/update/", RuleCreate.as_view(), name='rules-create'),
-    path("api/v1/rules/", RuleListView.as_view(), name='rules-list'),
+    path("api/v1/rules/", include('rule.urls')),
 ]
